account/urls.py
from django.urls import path,include
from . import views
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

urlpatterns = [
    path('api/login', TokenObtainPairView.as_view(), name='login_view'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/register',views.RegisterAPIView.as_view(), name='register_view'),
    path('api/logout',views.LogOutAPIView.as_view(), name='logout_view'),
]